@@ -1,32 +1,33 @@
 name: spotify_sdk
 description: A flutter plugin that let's you communicate with the spotify sdk and auth lib
-version: 2.1.0
+version: 3.0.0-dev.3
 homepage: https://github.com/brim-borium/spotify_sdk
 issue_tracker: https://github.com/brim-borium/spotify_sdk/issues
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
-  flutter: ">=1.12.0"
+  sdk: '>=3.1.1 <4.0.0'
+  flutter: ">=1.17.0"
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_web_plugins:
     sdk: flutter
-  logger: ^1.0.0
-  dio: ^4.0.0
-  json_annotation: ^4.0.0
-  js: ^0.6.3
-  crypto: ^3.0.1
-  synchronized: ^3.0.0
+  logger: ^2.0.2+1
+  dio: ^5.3.3
+  json_annotation: ^4.8.1
+  js: ">=0.6.7 <0.8.0"
+  crypto: ^3.0.3
+  synchronized: ^3.1.0
+  http: ^1.1.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^2.0.5
-  json_serializable: ^4.1.3
-  pedantic: ^1.11.1
-   
+  build_runner: ^2.4.6
+  json_serializable: ^6.7.1
+  flutter_lints: ^3.0.1
+
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -39,10 +40,10 @@ flutter:
   # adding or updating assets for this project.
   plugin:
     platforms:
-      android: 
+      android:
         package: de.minimalme.spotify_sdk
         pluginClass: SpotifySdkPlugin
-      ios: 
+      ios:
         package: de.minimalme.spotify_sdk
         pluginClass: SpotifySdkPlugin
       web:
