@@ -4,7 +4,7 @@ public class SpotifySdkConstants
 {
     //connecting
     public static let methodConnectToSpotify = "connectToSpotify"
-    public static let methodGetAuthenticationToken = "getAuthenticationToken"
+    public static let methodGetAccessToken = "getAccessToken"
     public static let methodDisconnectFromSpotify = "disconnectFromSpotify"
     public static let methodCheckIfSpotifyAppIsActive = "isSpotifyAppActive"
 
