@@ -2,7 +2,7 @@ import Flutter
 import SpotifyiOS
 
 public class SwiftSpotifySdkPlugin: NSObject, FlutterPlugin {
-
+    private static var instance = SwiftSpotifySdkPlugin()
     private var appRemote: SPTAppRemote?
     private var connectionStatusHandler: ConnectionStatusHandler?
     private var playerStateHandler: PlayerStateHandler?
@@ -11,17 +11,17 @@ public class SwiftSpotifySdkPlugin: NSObject, FlutterPlugin {
     private static var playerContextChannel: FlutterEventChannel?
 
     public static func register(with registrar: FlutterPluginRegistrar) {
+        guard playerStateChannel == nil else {
+            // Avoid multiple plugin registations
+            return
+        }
         let spotifySDKChannel = FlutterMethodChannel(name: "spotify_sdk", binaryMessenger: registrar.messenger())
         let connectionStatusChannel = FlutterEventChannel(name: "connection_status_subscription", binaryMessenger: registrar.messenger())
         playerStateChannel = FlutterEventChannel(name: "player_state_subscription", binaryMessenger: registrar.messenger())
         playerContextChannel = FlutterEventChannel(name: "player_context_subscription", binaryMessenger: registrar.messenger())
-
-        let instance = SwiftSpotifySdkPlugin()
         registrar.addApplicationDelegate(instance)
         registrar.addMethodCallDelegate(instance, channel: spotifySDKChannel)
-
         instance.connectionStatusHandler = ConnectionStatusHandler()
-
         connectionStatusChannel.setStreamHandler(instance.connectionStatusHandler)
     }
 
@@ -70,7 +70,7 @@ public class SwiftSpotifySdkPlugin: NSObject, FlutterPlugin {
                 return
             }
 
-        case SpotifySdkConstants.methodGetAuthenticationToken:
+        case SpotifySdkConstants.methodGetAccessToken:
             guard let swiftArguments = call.arguments as? [String:Any],
                 let clientID = swiftArguments[SpotifySdkConstants.paramClientId] as? String,
                 let url = swiftArguments[SpotifySdkConstants.paramRedirectUrl] as? String else {
