@@ -6,12 +6,10 @@ part of 'album.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-Album _$AlbumFromJson(Map<String, dynamic> json) {
-  return Album(
-    json['name'] as String,
-    json['uri'] as String,
-  );
-}
+Album _$AlbumFromJson(Map<String, dynamic> json) => Album(
+      json['name'] as String?,
+      json['uri'] as String?,
+    );
 
 Map<String, dynamic> _$AlbumToJson(Album instance) => <String, dynamic>{
       'name': instance.name,
