@@ -6,12 +6,10 @@ part of 'artist.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-Artist _$ArtistFromJson(Map<String, dynamic> json) {
-  return Artist(
-    json['name'] as String?,
-    json['uri'] as String?,
-  );
-}
+Artist _$ArtistFromJson(Map<String, dynamic> json) => Artist(
+      json['name'] as String?,
+      json['uri'] as String?,
+    );
 
 Map<String, dynamic> _$ArtistToJson(Artist instance) => <String, dynamic>{
       'name': instance.name,
