@@ -6,11 +6,9 @@ part of 'capabilities.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-Capabilities _$CapabilitiesFromJson(Map<String, dynamic> json) {
-  return Capabilities(
-    canPlayOnDemand: json['can_play_on_demand'] as bool,
-  );
-}
+Capabilities _$CapabilitiesFromJson(Map<String, dynamic> json) => Capabilities(
+      canPlayOnDemand: json['can_play_on_demand'] as bool,
+    );
 
 Map<String, dynamic> _$CapabilitiesToJson(Capabilities instance) =>
     <String, dynamic>{
