@@ -6,11 +6,9 @@ part of 'image_uri.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-ImageUri _$ImageUriFromJson(Map<String, dynamic> json) {
-  return ImageUri(
-    json['raw'] as String,
-  );
-}
+ImageUri _$ImageUriFromJson(Map<String, dynamic> json) => ImageUri(
+      json['raw'] as String,
+    );
 
 Map<String, dynamic> _$ImageUriToJson(ImageUri instance) => <String, dynamic>{
       'raw': instance.raw,
