@@ -6,22 +6,20 @@ part of 'track.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-Track _$TrackFromJson(Map<String, dynamic> json) {
-  return Track(
-    Album.fromJson(json['album'] as Map<String, dynamic>),
-    Artist.fromJson(json['artist'] as Map<String, dynamic>),
-    (json['artists'] as List<dynamic>)
-        .map((e) => Artist.fromJson(e as Map<String, dynamic>))
-        .toList(),
-    json['duration_ms'] as int,
-    ImageUri.fromJson(json['image_id'] as Map<String, dynamic>),
-    json['name'] as String,
-    json['uri'] as String,
-    json['linked_from_uri'] as String?,
-    isEpisode: json['is_episode'] as bool,
-    isPodcast: json['is_podcast'] as bool,
-  );
-}
+Track _$TrackFromJson(Map<String, dynamic> json) => Track(
+      Album.fromJson(json['album'] as Map<String, dynamic>),
+      Artist.fromJson(json['artist'] as Map<String, dynamic>),
+      (json['artists'] as List<dynamic>)
+          .map((e) => Artist.fromJson(e as Map<String, dynamic>))
+          .toList(),
+      json['duration_ms'] as int,
+      ImageUri.fromJson(json['image_id'] as Map<String, dynamic>),
+      json['name'] as String,
+      json['uri'] as String,
+      json['linked_from_uri'] as String?,
+      isEpisode: json['is_episode'] as bool,
+      isPodcast: json['is_podcast'] as bool,
+    );
 
 Map<String, dynamic> _$TrackToJson(Track instance) => <String, dynamic>{
       'album': instance.album,
