@@ -6,13 +6,11 @@ part of 'user_status.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-UserStatus _$UserStatusFromJson(Map<String, dynamic> json) {
-  return UserStatus(
-    json['code'] as int,
-    json['short_text'] as String,
-    json['long_text'] as String,
-  );
-}
+UserStatus _$UserStatusFromJson(Map<String, dynamic> json) => UserStatus(
+      json['code'] as int,
+      json['short_text'] as String,
+      json['long_text'] as String,
+    );
 
 Map<String, dynamic> _$UserStatusToJson(UserStatus instance) =>
     <String, dynamic>{
